move delete_orphans
