+++ cms/djangoapps/contentstore/management/commands/delete_orphans.py
@@ -5,7 +5,7 @@ from django.core.management.base import BaseCommand, CommandError
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey
 
-from cms.djangoapps.contentstore.xblock_storage_handlers.helpers import delete_orphans
+from cms.djangoapps.contentstore.xblock_storage_handlers import delete_orphans
 from xmodule.modulestore import ModuleStoreEnum  # lint-amnesty, pylint: disable=wrong-import-order
